# backend/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Basis Social API",
    "version": "1.0.0-alpha",
    "description": "Stable REST and world-server integration contract for BasisVR. ActivityPub endpoints are intentionally outside this client API contract."
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/api/v1/worlds": {
      "get": {"tags":["Worlds"],"operationId":"listWorlds","parameters":[{"$ref":"#/components/parameters/Limit"},{"$ref":"#/components/parameters/Cursor"}],"responses":{"200":{"description":"Cursor page of public worlds","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorldPage"}}}},"default":{"$ref":"#/components/responses/Error"}}},
      "post": {"tags":["Worlds"],"operationId":"createWorld","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorldWrite"}}}},"responses":{"201":{"description":"Created world","content":{"application/json":{"schema":{"$ref":"#/components/schemas/World"}}}},"default":{"$ref":"#/components/responses/Error"}}}
    },
    "/api/v1/worlds/{worldRef}": {"get": {"tags":["Worlds"],"operationId":"getWorld","parameters":[{"$ref":"#/components/parameters/WorldRef"}],"responses":{"200":{"description":"World; GET resolves worldRef as slug.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/World"}}}},"default":{"$ref":"#/components/responses/Error"}}},"patch": {"tags":["Worlds"],"operationId":"updateWorld","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/WorldRef"}],"responses":{"200":{"description":"Updated world; PATCH resolves worldRef as UUID."},"default":{"$ref":"#/components/responses/Error"}}},"delete":{"tags":["Worlds"],"operationId":"deleteWorld","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/WorldRef"}],"responses":{"204":{"description":"Deleted; DELETE resolves worldRef as UUID."},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/worlds/{id}/favorite": {"post": {"tags":["Worlds"],"operationId":"favoriteWorld","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"200":{"description":"Favorited"},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/worlds/{id}/announce": {"post": {"tags":["Worlds"],"operationId":"announceWorld","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"202":{"description":"Announcement accepted"},"default":{"$ref":"#/components/responses/Error"}}}},
...
    "/api/v1/groups/{id}/worlds": {"get":{"tags":["Groups"],"operationId":"listGroupWorlds","parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"200":{"description":"Linked worlds"},"default":{"$ref":"#/components/responses/Error"}}},"post":{"tags":["Groups"],"operationId":"addGroupWorld","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/ID"}],"responses":{"204":{"description":"World linked"},"default":{"$ref":"#/components/responses/Error"}}}},
    "/api/v1/groups/{id}/worlds/{worldId}": {"delete":{"tags":["Groups"],"operationId":"removeGroupWorld","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/ID"},{"name":"worldId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"204":{"description":"World unlinked"},"default":{"$ref":"#/components/responses/Error"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type":"http","scheme":"bearer","bearerFormat":"JWT"},
      "serviceToken": {"type":"apiKey","in":"header","name":"X-Basis-Service-Token"}
    },
    "parameters": {
      "ID": {"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},
...
      "WorldRef": {"name":"worldRef","in":"path","required":true,"schema":{"type":"string"}},
...
      "Limit": {"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"maximum":100,"default":24}},
      "Cursor": {"name":"cursor","in":"query","description":"Opaque cursor returned by the previous page.","schema":{"type":"string"}}
    },
    "responses": {
      "Error": {"description":"Error response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}
    },
...
    "schemas": {
      "Error": {"type":"object","required":["error"],"properties":{"error":{"type":"object","required":["code","message"],"properties":{"code":{"type":"string"},"message":{"type":"string"}}}}},
...
      "Owner": {"type":"object","required":["actorId","acct","displayName"],"properties":{"actorId":{"type":"string","format":"uuid"},"acct":{"type":"string"},"displayName":{"type":"string"}}},
      "Pagination": {"type":"object","required":["nextCursor","limit"],"properties":{"nextCursor":{"type":["string","null"]},"limit":{"type":"integer","minimum":1,"maximum":100}}},
      "World": {"type":"object","required":["id","slug","name","owner","visibility","capacity","tags","metadata"],"properties":{"id":{"type":"string","format":"uuid"},"slug":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"owner":{"$ref":"#/components/schemas/Owner"},"previewUrl":{"type":"string"},"launchUrl":{"type":"string"},"visibility":{"type":"string","enum":["public","followers","friends","private"]},"capacity":{"type":"integer","minimum":0},"tags":{"type":"array","items":{"type":"string"}},"metadata":{"type":"object"},"activityPub":{"type":"object"}}},
      "WorldWrite": {"type":"object","required":["name"],"properties":{"slug":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"previewUrl":{"type":"string"},"launchUrl":{"type":"string"},"visibility":{"type":"string"},"capacity":{"type":"integer","minimum":0},"tags":{"type":"array","items":{"type":"string"}},"metadata":{"type":"object"}}},
      "WorldPage": {"type":"object","required":["data","pagination"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/World"}},"pagination":{"$ref":"#/components/schemas/Pagination"}}},
...
    }
  }
}